1.4.4 initial commit
It builds!

Main things to do:
-Re-implement gravpatch for rod of discord aiming

-Re-implement the ancient dragon lance inheriting trident effects

-Ensure that changing CloneWithModdedData to Clone in EmeraldHeraldUI didn't break anything

-Fix SoulSlotRecipePatch IL edit

-Fix SpawnPatch method swap

+++ DarkSoulCustomCurrency.cs
@@ -11,7 +11,8 @@ namespace tsorcRevamp
         public DarkSoulCustomCurrency(int coinItemID, long currencyCap) : base(coinItemID, currencyCap)
         {
         }
-        public override void GetPriceText(string[] lines, ref int currentLine, int price)
+
+        public override void GetPriceText(string[] lines, ref int currentLine, long price)
         {
             lines[currentLine++] = string.Format("[c/00FF50:{0} {1} {2}]" + $"[i:{ModContent.ItemType<SoulCoin>()}]", new object[]
             {
